PSoC Workspace/GUI/requirements.txt
cycler==0.11.0
fonttools==4.28.2
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.3
packaging==21.3
Pillow==8.4.0
pygame==2.1.0
PyOpenGL==3.1.5
pyparsing==3.0.6
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.3
pyserial==3.5
PySide2==5.15.2
python-dateutil==2.8.2
setuptools-scm==6.3.2
shiboken2==5.15.2
six==1.16.0
tomli==1.2.2
